Update
You are developing a user registration system for a website. The username entered by the user must follow these rules:

It should contain only alphanumeric characters (a-z, A-Z, 0-9).
It should start with a letter (not a number or special character).
It should be at least 5 characters long and at most 15 characters long.